visualize one class

+++ main.py
@@ -159,7 +159,7 @@ aux_label = Variable(aux_label)
 # noise for evaluation
 eval_noise_ = np.random.normal(0, 1, (opt.batchSize, nz))
 if opt.visualize_class_label > 0:
-    eval_label = np.ones(opt.batchSize) * opt.visualize_class_label
+    eval_label = np.ones(opt.batchSize, dtype=np.int) * opt.visualize_class_label
 else:
     eval_label = np.random.randint(0, num_classes, opt.batchSize)
 eval_onehot = np.zeros((opt.batchSize, num_classes))
